add normal router.use(middleware) handling
