+++ browser.js
@@ -3,41 +3,7 @@ var pathToRegexp = require("path-to-regexp");
 var browserEnv   = require("./lib/dom_event_handler");
 var historyEnv   = require("./lib/history");
 var Qs           = require("qs");
-
-/**
- * make a url from an href or action attribute into a url isoRouter can handle
- *
- * If clicking a link on a page hosted at /land?animal=badger#grass
- *
- * If a host or pathname there is enough to route
- * /sea                    -> /sea
- * /sea#weed               -> /sea#weed
- * /sea?fish=flounder      -> /sea?fish=flounder
- * /sea?fish=flounder#weed -> /sea?fish=flounder#weed
- *
- * If only a search query is given it is intended to be relative to the current path
- * ?fish=flounder          -> /land?fish=flounder
- * ?fish=flounder#weed     -> /land?fish=flounder#weed
- *
- * If only a hash is given it is intended to be relative to the current search query
- * #sea                    -> /land?animal=badger#sea
- *
- * @param {String}  url     input url to normalize
- * @returns {String} normalized url ensuring hash, search and path are appropriate
- */
-function tidyUrl (url) {
-  var parsedUrl = Url.parse(url, true);
-
-  if (!parsedUrl.host && !parsedUrl.pathname) {
-    parsedUrl.pathname = window.location.pathname;
-
-    if (parsedUrl.hash && !parsedUrl.search) {
-      parsedUrl.search = window.location.search;
-    }
-  }
-
-  return parsedUrl;
-}
+var urlParser      = require("./lib/url_parser");
 
 /**
  * handler - create a route handler
@@ -48,9 +14,22 @@ function tidyUrl (url) {
  * @returns {undefined}
  */
 function addRouteHandler (method, path, fn) {
+
+  // If using router.use(middleware) then set a catch all path
+  if (!fn) {
+    fn = path;
+    path = "/*";
+  }
+
   var re = pathToRegexp(path);
   var keys = re.keys;
 
+  /**
+   * Construct a routeHandler
+   * @param {String} _method     HTTP method of the incoming request
+   * @param {String} _path       url path of the incoming request
+   * @returns {false|Function} express like middleware
+   */
   function routeHandler (_method, _path) {
     // Check the method matches
     if (method !== "use" && method !== _method) {
@@ -64,6 +43,14 @@ function addRouteHandler (method, path, fn) {
       return false;
     }
 
+    /**
+     * Wrap a route handler
+     * @param {Error} [err]        optional error denoting should pass request to error handling middleware
+     * @param {Object} req         request object
+     * @param {Object} res         response object
+     * @param {Function} next      continuation function called with an optional error param
+     * @returns {Void} next will be called
+     */
     return function handler () {
       var err, req, res, next;
       if (arguments.length === 4) {
@@ -88,18 +75,16 @@ function addRouteHandler (method, path, fn) {
         if (fn.length < 4) {
           next(err);
         } else {
-          return fn(err, req, res, next);
+          fn(err, req, res, next);
         }
       } else {
-        return fn(req, res, next);
+        fn(req, res, next);
       }
     };
   }
 
   this.routes.push(routeHandler);
 }
-
-var uid = 0;
 
 /**
  * go - peform a request to router
@@ -123,7 +108,7 @@ function go (path, opts) {
   var body = opts.body || {};
   var locals = opts.locals || {};
 
-  var parsedUrl = tidyUrl(path);
+  var parsedUrl = urlParser(path);
   var url = parsedUrl.format(parsedUrl);
 
   // If redirecting to self catch and do a full navigation
@@ -145,7 +130,7 @@ function go (path, opts) {
 
   // request object, similar to express
   var req = {
-    __id: uid++,
+    __id: this.reqIdx++,
     body: body,
     locals: locals,
     originalUrl: url,
@@ -253,9 +238,10 @@ module.exports = function clientRouter (opts) {
 
   // Setup the context
   var ctx = {
-    routes: [],
-    history: [],
-    selfRedirectCount: 0
+    routes: [], // Array of route handlers to be run through on each request
+    history: [], // Array of objects containing request method and path
+    selfRedirectCount: 0, // Count of how many times the same url has been hit
+    reqIdx: 0 // incrementing count of each request
   };
 
   function removeDomEventHandler () {
@@ -263,14 +249,6 @@ module.exports = function clientRouter (opts) {
       domEventHandler.destroy();
     }
   }
-
-  // When the url changes (such as back button) want to trigger the appropriate handler
-  window.addEventListener("popstate", function () {
-    var url = document.location.pathname + document.location.search;
-    go.call(ctx, url, {
-      silent: true
-    });
-  });
 
   // Expose the router API
   router = {
@@ -283,6 +261,14 @@ module.exports = function clientRouter (opts) {
     removeDomEventHandler: removeDomEventHandler,
     history: historyEnv
   };
+
+  // When the url changes (such as back button) want to trigger the appropriate handler
+  window.addEventListener("popstate", function () {
+    var url = document.location.pathname + document.location.search;
+    router.go(url, {
+      silent: true
+    });
+  });
 
   // Injects a delegate event listener onto window or a specific node
   if (opts.inject) {
